backend/requirements-gpu-directml.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-hub==0.12.0
keras==2.10.0
opencv-python==4.9.0.80
moviepy==1.0.3
numpy==1.23.5
pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
matplotlib==3.8.4
reportlab==4.2.0
python-dotenv==1.0.1
waitress==3.0.0
gunicorn==22.0.0
psutil==5.9.8
imageio==2.34.1
imageio-ffmpeg==0.5.1
requests==2.33.1
